fix(ipi): small fixes to the ipi implementation
- Fix the return type for the api_hf_interrupt_send_ipi function. In
the case the target vcpu id is not valid it returns -1 so the type
must be signed.
- Fix the cast in the hf_call in hf_interrupt_send_ipi, the arguments
for this function are uint64_t so cast to this.
- Move hf_ipi_init_interrupt to arch_cpu_init as this is a more
appropriate place for it.
- Lock the cpu when setting and getting the target vCPU with
pending IPI.
- Set the IPI and SRI interrupts to enabled during initialization.

+++ inc/vmapi/hf/call.h
@@ -412,7 +412,7 @@ static inline int64_t hf_interrupt_reconfigure_target_cpu(
  */
 static inline int64_t hf_interrupt_send_ipi(ffa_vcpu_index_t target_vcpu_id)
 {
-	return hf_call(HF_INTERRUPT_SEND_IPI, (uint32_t)target_vcpu_id, 0, 0);
+	return hf_call(HF_INTERRUPT_SEND_IPI, (uint64_t)target_vcpu_id, 0, 0);
 }
 
 /** Obtains the Hafnium's version of the implemented FF-A specification. */
